bcmaps: Map Layers and Spatial Utilities for British Columbia

Various layers of B.C., including administrative boundaries, natural resource management boundaries, census boundaries etc. All layers are available in BC Albers (https://spatialreference.org/ref/epsg/3005/) equal-area projection, which is the B.C. government standard. The layers are sourced from the British Columbia and Canadian government under open licenses, including B.C. Data Catalogue (https://data.gov.bc.ca), the Government of Canada Open Data Portal (https://open.canada.ca/en/using-open-data), and Statistics Canada (https://www.statcan.gc.ca/en/reference/licence).

See also

Author

Maintainer: Andy Teucher andy.teucher@gmail.com (ORCID)

Authors:

Sam Albers sam.albers@gmail.com (ORCID) [contributor]

Stephanie Hazlitt stephhazlitt@gmail.com (ORCID) [contributor]

Other contributors:

Province of British Columbia [copyright holder]